# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns current user profile"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/request-verification": {
      "post": {
        "operationId": "UsersController_requestVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUserVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification request submitted successfully"
          },
          "400": {
            "description": "Already verified or verification already pending"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request user verification (submit identity document and selfie)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/pending-verifications": {
      "get": {
        "operationId": "UsersController_getPendingVerifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns list of users pending verification"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Moderator] Get all users with pending verification",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/verification-details": {
      "get": {
        "operationId": "UsersController_getVerificationDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user verification details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Moderator] Get user verification details including documents",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/verification-status": {
      "patch": {
        "operationId": "UsersController_updateVerificationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification status updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Moderator] Approve or reject user verification",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Transparent Donation System API",
    "description": "\n## Overview\nA blockchain-enabled donation platform ensuring transparency, auditability, and immutability.\n\n## Features\n- **User Management**: Registration, authentication, and verification\n- **Organization Management**: CRUD with blockchain recording\n- **Campaign Management**: Create and manage fundraising campaigns\n- **Donations**: Anonymous donation support with blockchain verification\n- **Verification System**: Multi-level verification for users, organizations, and campaigns\n- **Audit Trail**: Complete blockchain-backed audit history\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Roles\n- **USER**: Standard user account\n- **MODERATOR**: Reviews and approves verification requests\n- **AUDITOR**: Read-only access to verification and audit data\n- **ADMIN**: System administration and operational exception handling\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User full name"
          },
          "password": {
            "type": "string",
            "description": "New password",
            "minLength": 6
          },
          "phone": {
            "type": "string",
            "description": "User phone number",
            "example": "0123456789"
          },
          "address": {
            "type": "string",
            "description": "User address",
            "example": "Hà Nội"
          }
        }
      },
      "RequestUserVerificationDto": {
        "type": "object",
        "properties": {
          "identityDocument": {
            "type": "string",
            "description": "Identity document URL (CCCD/Passport image)",
            "example": "https://storage.example.com/docs/cccd.jpg"
          },
          "selfieWithDocument": {
            "type": "string",
            "description": "Selfie photo URL holding the identity document",
            "example": "https://storage.example.com/docs/selfie-with-cccd.jpg"
          },
          "verificationNote": {
            "type": "string",
            "description": "Additional note for verification (optional)",
            "example": "Thông tin bổ sung nếu cần"
          }
        },
        "required": [
          "identityDocument",
          "selfieWithDocument"
        ]
      },
...
    }
  }
}